+++ src/revan/inc/MFileEventsEvta.h
@@ -71,7 +71,7 @@ class MFileEventsEvta : public MFileEvents
   MGeometryRevan* m_Geometry;
   MString m_GeometryFileName;
 
-  int m_EventId;
+  long m_EventId;
   bool m_IsSimulation;
   bool m_IsFirstEvent;
 
@@ -81,7 +81,7 @@ class MFileEventsEvta : public MFileEvents
   //! The ER responsible for noising the data
   MERNoising* m_Noising;
   
-  static const int c_NoId;
+  static const long c_NoId;
 
 
 #ifdef ___CINT___
